+++ packages/client/src/pages/CarbonHackCards/ForecastCard/ForecastCard.tsx
@@ -1,14 +1,19 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import React, { FunctionComponent, ReactElement } from 'react'
-import { useForecastData } from '../../../utils/hooks/ForecastDataHook'
+import { useForecastData } from 'src/utils/hooks/ForecastDataHook'
 import useStyles from '../../EmissionsMetricsPage/CarbonComparisonCard/carbonComparisonStyles'
 import DashboardCard from '../../../layout/DashboardCard'
+import Table from '@mui/material/Table'
+import TableBody from '@mui/material/TableBody'
+import TableCell from '@mui/material/TableCell'
+import TableContainer from '@mui/material/TableContainer'
+import TableHead from '@mui/material/TableHead'
+import TableRow from '@mui/material/TableRow'
+import Paper from '@mui/material/Paper'
 import { Typography, CardContent } from '@material-ui/core'
 import moment from 'moment'
-import LoadingMessage from '../../../common/LoadingMessage'
-import { DataGrid, GridColumns, GridOverlay } from '@mui/x-data-grid'
 import LineChartDialog from './LineChartDialog'
-import ErrorPage from '../../../layout/ErrorPage'
-
 const ForecastCard: FunctionComponent<any> = ({ data }): ReactElement => {
   const convertUTCtoLocalTime = (utcTime) => {
     const stillUtc = moment.utc(utcTime).toDate()
@@ -28,107 +33,43 @@ const ForecastCard: FunctionComponent<any> = ({ data }): ReactElement => {
   })
 
   const optimalRegionTimeMap = (result) => {
-    const row = []
+    // eslint-disable-next-line prefer-const
+    let array = []
     for (const [key, value] of result) {
       if (value != undefined) {
-        value.optimalDataPoints.map((eachOptimalTime) => {
+        value.optimalDataPoints.map((eachOptimalTime, index) => {
           const [date, time] = convertUTCtoLocalTime(eachOptimalTime.timestamp)
-          row.push({
-            id: key,
-            region: accountRegionMap.get(key)[0],
-            date: date,
-            time: time,
-            rating: eachOptimalTime.value,
-            moreinfo: {
-              forecastData: value.forecastData,
-              region: accountRegionMap.get(key)[0],
-            },
-          })
+          array.push(
+            <>
+              <TableRow
+                key={index}
+                sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
+              >
+                <TableCell align="center">{key}</TableCell>
+                <TableCell align="center">
+                  {accountRegionMap.get(key)[0]}
+                </TableCell>
+                <TableCell align="center">{date}</TableCell>
+                <TableCell align="center">{time}</TableCell>
+                <TableCell align="center">{eachOptimalTime.value}</TableCell>
+                <TableCell align="center">
+                  <LineChartDialog
+                    forecastData={value.forecastData}
+                    region={accountRegionMap.get(key)[0]}
+                  ></LineChartDialog>
+                </TableCell>
+              </TableRow>
+            </>,
+          )
         })
       }
     }
-    return row
+    return array
   }
 
   const dataToSend = accountRegionMap
 
-  const { result, error, loading } = useForecastData(dataToSend)
-
-  const columns: GridColumns = [
-    {
-      field: 'id',
-      headerName: 'Account',
-      width: 250,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'region',
-      headerName: 'Region',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'date',
-      headerName: 'Date',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'time',
-      headerName: 'Time',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'rating',
-      headerName: 'Rating (g/kWh)',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'moreinfo',
-      headerName: 'More Info',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-      renderCell: (params) => (
-        <LineChartDialog
-          testId="line-chart-dialog"
-          forecastData={params.value.forecastData}
-          region={params.value.region}
-        ></LineChartDialog>
-      ),
-    },
-  ]
-
-  const renderTable = () => {
-    return (
-      <div style={{ height: 371, width: '100%' }}>
-        <DataGrid
-          rows={optimalRegionTimeMap(result)}
-          columns={columns}
-          pageSize={5}
-          rowsPerPageOptions={[5]}
-          columnBuffer={6}
-          rowBuffer={5}
-          hideFooterSelectedRowCount={true}
-          components={{
-            NoRowsOverlay: () => (
-              <GridOverlay>
-                There's no data to display! Expand your search parameters to get
-                started. (Try adding accounts, regions or recommendation types)
-              </GridOverlay>
-            ),
-          }}
-        />
-      </div>
-    )
-  }
+  const result = useForecastData(dataToSend)
 
   return (
     <DashboardCard noPadding>
@@ -147,19 +88,21 @@ const ForecastCard: FunctionComponent<any> = ({ data }): ReactElement => {
           </Typography>
           <Typography className={classes.posOne}></Typography>
         </CardContent>
-        {error != null ? (
-          <ErrorPage errorMessage={'Error loading cloud data'} />
-        ) : (
-          <>
-            {loading ? (
-              <LoadingMessage
-                message={'Loading cloud data. This may take a while...'}
-              />
-            ) : (
-              renderTable()
-            )}
-          </>
-        )}
+        <TableContainer component={Paper}>
+          <Table sx={{ minWidth: 650 }} aria-label="simple table">
+            <TableHead>
+              <TableRow>
+                <TableCell align="center">Account</TableCell>
+                <TableCell align="center">Region</TableCell>
+                <TableCell align="center">Date</TableCell>
+                <TableCell align="center">Time</TableCell>
+                <TableCell align="center">Rating (g/kWh)</TableCell>
+                <TableCell align="center">More Info</TableCell>
+              </TableRow>
+            </TableHead>
+            <TableBody>{optimalRegionTimeMap(result)}</TableBody>
+          </Table>
+        </TableContainer>
       </React.Fragment>
     </DashboardCard>
   )
