+++ packages/client/src/pages/CarbonHackCards/RegionRecommendationCard/RegionRecommendationCard.tsx
@@ -1,21 +1,24 @@
 import React, { FunctionComponent, ReactElement } from 'react'
 import { useRegionRecommendationData } from '../../../utils/hooks/RegionRecommendationDataHook'
 import useStyles from '../../EmissionsMetricsPage/CarbonComparisonCard/carbonComparisonStyles'
-import { AWS_REGIONS, AZURE_REGIONS, GCP_REGIONS } from './CloudRegions'
+import { AWS_REGIONS } from './AWSRegions'
 import DashboardCard from '../../../layout/DashboardCard'
+import Table from '@mui/material/Table'
+import TableBody from '@mui/material/TableBody'
+import TableCell from '@mui/material/TableCell'
+import TableContainer from '@mui/material/TableContainer'
+import TableHead from '@mui/material/TableHead'
+import TableRow from '@mui/material/TableRow'
+import Paper from '@mui/material/Paper'
 import { Typography, CardContent } from '@material-ui/core'
-import { DataGrid, GridColumns, GridOverlay } from '@mui/x-data-grid'
-import LoadingMessage from '../../../common/LoadingMessage'
-import ErrorPage from '../../../layout/ErrorPage'
-
 const RegionRecommendationCard: FunctionComponent<any> = ({
   data,
 }): ReactElement => {
   const accountRegionMap = new Map()
 
   const classes = useStyles()
-
   const footPrintData = data !== undefined ? data : []
+  console.log('Footprint: ', footPrintData)
 
   const findNearestRegions = (
     accountRegion: string,
@@ -24,29 +27,22 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
     if (cloudProvider === 'AWS') {
       return getNearestRegionFromEnum(accountRegion, AWS_REGIONS)
     } else if (cloudProvider === 'AZURE') {
-      return getNearestRegionFromObject(accountRegion, AZURE_REGIONS)
+      return ['ukwest', 'UK South']
     } else {
-      return getNearestRegionFromEnum(accountRegion, GCP_REGIONS)
+      return ['us-east1', 'us-west1', 'us-east4', 'us-west2']
     }
+  }
+
+  const truncateDecimals = function (number) {
+    return Math.floor(number)
   }
 
   const getNearestRegionFromEnum = (accountRegion: string, REGIONS: any) => {
     const nearestRegions = []
     const countryPrefix: string = accountRegion.slice(0, 2).toUpperCase()
-    for (const regionName in REGIONS) {
-      if (regionName.startsWith(countryPrefix)) {
-        nearestRegions.push(REGIONS[regionName])
-      }
-    }
-    return nearestRegions
-  }
-
-  const getNearestRegionFromObject = (accountRegion: string, REGIONS: any) => {
-    const nearestRegions = []
-    const countryPrefix: string = accountRegion.slice(0, 2).toUpperCase()
-    for (const regionPair in Object.entries(REGIONS)) {
-      if (regionPair[0].startsWith(countryPrefix)) {
-        nearestRegions.push(regionPair[1])
+    for (const awsRegionName in REGIONS) {
+      if (awsRegionName.startsWith(countryPrefix)) {
+        nearestRegions.push(REGIONS[awsRegionName])
       }
     }
     return nearestRegions
@@ -54,88 +50,23 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
 
   footPrintData.map((month) => {
     month.serviceEstimates.map((account) => {
-      const date = new Date(
-        new Date(month.timestamp).setMonth(
-          new Date(month.timestamp).getMonth() + 1,
-        ),
-      )
-      const endMonth =
-        date.getMonth() < 9 ? '0' + (date.getMonth() + 1) : date.getMonth() + 1
-      if (accountRegionMap.get(account.accountName) != undefined) {
-        const previousCo2e = accountRegionMap.get(account.accountName)[2]
-        const previouskilowattHours = accountRegionMap.get(
-          account.accountName,
-        )[3]
-        accountRegionMap.set(account.accountName, [
-          account.region,
-          account.cloudProvider,
-          account.co2e + previousCo2e,
-          account.kilowattHours + previouskilowattHours,
-          findNearestRegions(account.region, account.cloudProvider),
-          month.timestamp,
-          `${date.getFullYear()}-${endMonth}-${date.getDate()}T00:00:00.000Z`,
-        ])
-      } else {
-        accountRegionMap.set(account.accountName, [
-          account.region,
-          account.cloudProvider,
-          account.co2e,
-          account.kilowattHours,
-          findNearestRegions(account.region, account.cloudProvider),
-          month.timestamp,
-          `${date.getFullYear()}-${endMonth}-${date.getDate()}T00:00:00.000Z`,
-        ])
-      }
+      accountRegionMap.set(account.accountName, [
+        account.region,
+        account.cloudProvider,
+        account.co2e,
+        account.kilowattHours,
+        findNearestRegions(account.region, account.cloudProvider),
+        month.timestamp,
+        '2020-02-25',
+      ])
     })
   })
 
-  const { result, error, loading } =
-    useRegionRecommendationData(accountRegionMap)
+  const dataToSend = accountRegionMap
+  console.log(dataToSend, ' in emussins card')
 
-  const columns: GridColumns = [
-    {
-      field: 'id',
-      headerName: 'Account',
-      width: 250,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'region',
-      headerName: 'Region',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'actualEmissions',
-      headerName: 'Actual Emisssions (Mg)',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'bestLocation',
-      flex: 0.75,
-      headerName: 'Best Location',
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'expectedEmissions',
-      headerName: 'Expected Emissions (Mg)',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'reduction',
-      headerName: 'Reduction in Emissions',
-      flex: 0.75,
-      headerAlign: 'center',
-      align: 'center',
-    },
-  ]
+  const result = useRegionRecommendationData(dataToSend)
+  console.log(result, 'result of emission card')
 
   const bestLocationTable = (result) => {
     // eslint-disable-next-line prefer-const
@@ -143,57 +74,41 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
     for (const [key, value] of result) {
       if (value != undefined) {
         const expectedEmissionsForBestLocation =
-          (value.carbonIntensity * accountRegionMap.get(key)[3]) / 1000000
+          (truncateDecimals(value.carbonIntensity) *
+            truncateDecimals(accountRegionMap.get(key)[3])) /
+          1000000
         const actualEmissions = accountRegionMap.get(key)[2]
-        const percentReduction =
-          actualEmissions === 0
-            ? 0
-            : 100 - (expectedEmissionsForBestLocation * 100) / actualEmissions
-        const reduction =
-          accountRegionMap.get(key)[0] === value.location
-            ? 0
-            : Math.floor(percentReduction)
-        const expectedEmissions =
-          accountRegionMap.get(key)[0] === value.location
-            ? actualEmissions.toFixed(5)
-            : expectedEmissionsForBestLocation.toFixed(5)
-        const bestLocation =
-          actualEmissions <= expectedEmissions ? '-' : value.location
-        array.push({
-          id: key,
-          region: accountRegionMap.get(key)[0],
-          actualEmissions: actualEmissions.toFixed(5),
-          bestLocation: bestLocation,
-          expectedEmissions: expectedEmissions,
-          reduction: `${reduction} %`,
-        })
+        const percentReduction = truncateDecimals(
+          100 - (expectedEmissionsForBestLocation * 100) / actualEmissions,
+        )
+        array.push(
+          <>
+            <TableRow
+              sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
+            >
+              <TableCell align="center">{key}</TableCell>
+              <TableCell component="th" scope="row" align="center">
+                {accountRegionMap.get(key)[0]}
+              </TableCell>
+              <TableCell align="center">{actualEmissions}</TableCell>
+              <TableCell align="center">{value.location}</TableCell>
+              <TableCell align="center">
+                {accountRegionMap.get(key)[0] === value.location
+                  ? actualEmissions
+                  : expectedEmissionsForBestLocation}
+              </TableCell>
+              <TableCell align="center">
+                {accountRegionMap.get(key)[0] === value.location
+                  ? 0
+                  : percentReduction}{' '}
+                %
+              </TableCell>
+            </TableRow>
+          </>,
+        )
       }
     }
     return array
-  }
-
-  const renderTable = () => {
-    return (
-      <div style={{ height: 371, width: '100%' }}>
-        <DataGrid
-          rows={bestLocationTable(result)}
-          columns={columns}
-          pageSize={5}
-          rowsPerPageOptions={[5]}
-          columnBuffer={6}
-          rowBuffer={5}
-          hideFooterSelectedRowCount={true}
-          components={{
-            NoRowsOverlay: () => (
-              <GridOverlay>
-                There's no data to display! Expand your search parameters to get
-                started. (Try adding accounts, regions or recommendation types)
-              </GridOverlay>
-            ),
-          }}
-        />
-      </div>
-    )
   }
 
   return (
@@ -209,23 +124,25 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
             data-testid="co2"
           >
             {' '}
-            Best Location to Shift your Deployment
+            Best Location to Shift your Deployed Location
           </Typography>
           <Typography className={classes.posOne}></Typography>
         </CardContent>
-        {error != null ? (
-          <ErrorPage errorMessage={'Error loading cloud data'} />
-        ) : (
-          <>
-            {loading ? (
-              <LoadingMessage
-                message={'Loading cloud data. This may take a while...'}
-              />
-            ) : (
-              renderTable()
-            )}
-          </>
-        )}
+        <TableContainer component={Paper}>
+          <Table sx={{ minWidth: 650 }} aria-label="simple table">
+            <TableHead>
+              <TableRow>
+                <TableCell align="center">Account</TableCell>
+                <TableCell align="center">Region</TableCell>
+                <TableCell align="center">Actual Emissions (Mg)</TableCell>
+                <TableCell align="center">Best Location</TableCell>
+                <TableCell align="center">Expected Emissions (Mg)</TableCell>
+                <TableCell align="center">Reduction in Emissions</TableCell>
+              </TableRow>
+            </TableHead>
+            <TableBody>{bestLocationTable(result)}</TableBody>
+          </Table>
+        </TableContainer>
       </React.Fragment>
     </DashboardCard>
   )
